scooby: Preserve whitespace in public/private notes.
The data from the previous input has newlines in it, which we do not
transform into `<br />` or anything; because of the rules of HTML,
this folds all whitespace upon display, merging any previous
multi-line public or private notes into one line.

Set these `white-space: pre` such that the newlines and potentially
any leading whitespace are faithfully rendered.  This makes them
round-trip correctly.