submitReport: Flag some tags if internal notes are unchanged.
Some tags are somewhat uncommon, but potentially reasonable if they
come with an updated internal notes field.  We could do this
server-side by making a round-trip to Airtable to check the previous
value of the internal notes, but this would add additional latency.

Instead, also submit the previous value of the internal notes from the
client, so we can determine if it has changed.

In general, this practice is *unsafe* because the client cannot be
trusted to not tamper with the values.  In this context, the client is
most probably not adversarial, and the only risk is that a malicious
user can cause a record to be potentially not marked for review with
these tags when they did not actually change the internal notes.  This
seems a low risk for the cost of saving the latency of an additional
round-trip to Airtable.

A more proper change would be to HMAC-sign the string when originally
providing internal notes from the server with some secret (say, the
Airtable token).  This is left as an exercise to the future
more-security-paranoid reader.