scooby: Link to the more general Walmart COVID vaccine page.
This page is more of a "landing" page.  The previous link is the more
direct link -- it takes you where the "Schedule now" button on the
landing page points to -- but it can be confusing to be dumped into a
contextless login page, with no hint that this is, indeed, for COVID
vaccines.

While it requires one more click, it likely reduces the bounce rate to
provide users with the necessary context.