Switch internal naming to match button.
In VIAL, `Yes: walk-ins accepted` was the "previous name" of a tag
that showed as "walk-ins accepted" and whose slug was
`walk_ins_accepted`.

CAVaccineInventory/vial#588 changed its internal name to be `Walk-ins
only` and its slug to be `walk_ins_only`.  Because Scooby submits
"Yes: walk-ins accepted", it can continue to work against new VIAL
code with no code changes.  However, some cleanup here will also
straighten out the confusion.

Submit `Walk-ins only`, (note the lack of leading `Yes: `) as the
availability tag, which is the *true name* of the tag, and not the
"previous name" that had been being used.  Also rename internal
variables to match.

This requires that CAVaccineInventory/vial#588 be deployed to work --
but this commit need not be deployed to straighten out the VIAL side.

Ref CAVaccineInventory/vial#587.